Include the Python API doc in the web site build
This has the effect of including the Python API static site into the
global web site.

This works in deployment as well, because the Python API static files
are not committed, nor are they known to the Git repository, hence they
are being included in the build _and_deployed.
See https://www.mkdocs.org/user-guide/deploying-your-docs/.