Fix samples version variable name typo on current site version build
Note: the site contents is correct because of the default values of
KHIOPS_SAMPLES_VERSION in mkdocs.yml for both `dev-v10` and `dev`
branches, which apply to the current versions of the respective
branches.

However, the samples version was incorrectly initialized upon calling
the action for building the current version of the site.